More flexible proxying
Adds an option to exit after forwarding an EAP success packet, to avoid causing
extra load by leaving interfaces in promiscuous mode. This is also useful in
case the authenticating client decides to send logoff messages later on.

Adds option to maintain a pidfile while running, which makes it easier to
guarantee that only one instance is started (useful for monitoring scripts).